+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added `Transform` properties and `Truncate` transformation
+
 ### Changed
 
 ### Fixed